backend/requirements.txt
web3==6.15.1
python-dotenv==1.0.0
supabase==2.3.4
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
